refactor: 分离会员积分相关模型到独立文件
